refactor: Move code to get urls into its own module
